requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

numpy>=1.23.5,<2
typing-extensions>=4.8.0
opencv-python==4.10.0.84
cv2_enumerate_cameras==1.1.15
onnx==1.18.0
insightface==0.7.3
psutil==5.9.8
tk==0.1.0
customtkinter==5.2.2
pillow==11.1.0
torch; sys_platform != 'darwin'
torch==2.8.0+cu128; sys_platform == 'darwin'
torchvision; sys_platform != 'darwin'
torchvision==0.20.1; sys_platform == 'darwin'
onnxruntime-silicon==1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime-gpu==1.22.0; sys_platform != 'darwin'
tensorflow; sys_platform != 'darwin'
opennsfw2==0.10.2
protobuf==4.25.1
git+https://github.com/xinntao/BasicSR.git@master
git+https://github.com/TencentARC/GFPGAN.git@master
pygrabber
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.19

# SaaS dependencies
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
authlib==1.3.2
python-jose[cryptography]==3.3.0
stripe==11.4.1
pydantic-settings==2.6.1
itsdangerous==2.2.0
